requirements.txt
Django>=3.2,<5.0
djangorestframework
django-cors-headers
gunicorn
whitenoise
psycopg2-binary==2.9.9
dj-database-url==0.5.0
